test: add failing test to cover nested objects for #356
